refactor: remove unused parameter and function, reorganize function
