+++ fwoa-tools/src/migrationVerify.ts
@@ -5,14 +5,14 @@
 import { readFileSync, WriteStream, createWriteStream } from 'fs';
 import { S3 } from 'aws-sdk';
 import { aws4Interceptor } from 'aws4-axios';
-import axios from 'axios';
+import axios, { AxiosInstance } from 'axios';
 import * as dotenv from 'dotenv';
+import objectHash from 'object-hash';
 import yargs from 'yargs';
 import {
   ExportOutput,
   getFhirClient,
   getFhirClientSMART,
-  verifyResource,
   checkConfiguration,
   EXPORT_STATE_FILE_NAME
 } from './migrationUtils';
@@ -46,6 +46,23 @@ function parseCmdOptions(): any {
 const argv: any = parseCmdOptions();
 const smartClient: boolean = argv.smart;
 const dryRun: boolean = argv.dryRun;
+
+async function verifyResource(
+  fhirClient: AxiosInstance,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  healthLakeResource: any,
+  resourceId: string,
+  resourceType: string
+): Promise<boolean> {
+  const fwoaResponse = (await fhirClient.get(`/${resourceType}/${resourceId}`)).data;
+  delete fwoaResponse.meta;
+  delete healthLakeResource.meta;
+  if (resourceType === 'Binary') {
+    delete healthLakeResource.data;
+    delete fwoaResponse.presignedGetUrl;
+  }
+  return objectHash(fwoaResponse) === objectHash(healthLakeResource);
+}
 async function verifyFolderImport(): Promise<void> {
   // eslint-disable-next-line security/detect-non-literal-fs-filename
   const outputFile: ExportOutput = JSON.parse(readFileSync(EXPORT_STATE_FILE_NAME).toString());
